Exports lined up
