+++ packages/frontend/src/hells-cubes/cockatrice/toCockCube.ts
@@ -6,6 +6,7 @@ import { recursiveAdoption } from '../recursiveAdoption.ts';
 import { prettifyXml } from './prettifyXml';
 import { getSplitSet } from '../../hellfall/filters/filterSet.ts';
 import namesRawData from '@hellfall/shared/data/oracle-names.json';
+import { mergeHCCardFaces } from '../mergeHCCardFaces';
 
 const hcToCockLayout: Record<HCLayout, string> = {
   normal: 'normal',
@@ -53,9 +54,9 @@ const hcToCockLayout: Record<HCLayout, string> = {
   vanguard: 'normal',
   station: 'normal',
 };
-const subLayouts = ['token'];
-const dropLayouts = ['meld_result', 'draft_partner', 'specialize'];
-const alwaysCompressLayouts = ['split', 'aftermath', 'prepare', 'adventure', 'token'];
+// const subLayouts = ['token'];
+// const dropLayouts = ['meld_result', 'draft_partner', 'specialize'];
+// const alwaysCompressLayouts = ['split', 'aftermath', 'prepare', 'adventure', 'token'];
 
 type CockFaceProps = Record<string, string | number | HCColors> & {
   name: string;
@@ -101,20 +102,19 @@ export const toCockCube = ({
   allCards: HCCard.Any[];
 }) => {
   const idNames: Record<string, string[]> = {};
-  const oracleNames = namesRawData.data;
   /**
    * Checks whether a name is taken
    * @param name name to check
    * @returns whether name is taken
    */
-  const nameIsTaken = (name: string): boolean => {
-    return oracleNames.includes(name) || Object.values(idNames).some(names => names.includes(name));
-  };
+  // const nameIsTaken = (name: string): boolean => {
+  //   return oracleNames.includes(name) || Object.values(idNames).some(names => names.includes(name));
+  // };
   const hcFaceToCockProps = (
     face: HCCard.AnySingleFaced | HCCardFace.MultiFaced
   ): CockFaceProps => {
     const cockFace: CockFaceProps = {
-      name: toExportName(face.name),
+      name: face.export_name || face.name,
       text: face.oracle_text.replaceAll(/\\n/g, '\n').replaceAll(/\{(.)\}/g, '$1'),
       layout: hcToCockLayout[face.layout],
       type: face.type_line,
@@ -122,10 +122,6 @@ export const toCockCube = ({
       manacost: face.mana_cost,
       cmc: face.mana_value,
     };
-    if ('id' in face && face.isActualToken && 'keywords' in face) {
-      // for tokens, use their unique id as the name to prevent conflicts
-      cockFace.name = face.id;
-    }
     if (face.colors.length) {
       cockFace.colors = orderColors(face.colors);
     }
@@ -133,7 +129,28 @@ export const toCockCube = ({
       cockFace.picurl = face.image;
     }
     if (face.power || face.toughness) {
-      cockFace.pt = (face.power || '') + '/' + (face.toughness || '');
+      const [powers,toughnesses] = [face.power?.split(' // '),face.toughness?.split(' // ')]
+
+      cockFace.pt = (powers?.[0] || '') + '/' + (toughnesses?.[0] || '');
+      if (powers && toughnesses) {
+        if (powers.length>1 || toughnesses.length > 1) {
+          for (let i = 1;i<Math.max(powers.length,toughnesses.length);i++) {
+            cockFace.pt += ` // ${powers[i] || ''}/${toughnesses[i] || ''}`;
+          }
+        }
+      } else if (powers) {
+        if (powers.length>1) {
+          for (let i = 1;i<powers.length;i++) {
+            cockFace.pt += ` // ${powers[i] || ''}/`;
+          }
+        }
+      } else {
+        if (toughnesses!.length>1) {
+          for (let i = 1;i<toughnesses!.length;i++) {
+            cockFace.pt += ` // /${toughnesses![i] || ''}`;
+          }
+        }
+      }
     }
     if (face.loyalty || face.defense) {
       cockFace.loyalty = face.loyalty ? face.loyalty : face.defense;
@@ -146,53 +163,53 @@ export const toCockCube = ({
    * @param goingBackwards true when order needs to be flipped (currently only bubsy uses this)
    * @returns merged cock face props
    */
-  const mergeCockFaceProps = (faces: CockFaceProps[], goingBackwards?: boolean): CockFaceProps => {
-    faces.slice(1).forEach((face, i) => {
-      Object.keys(face).forEach(key => {
-        if (key == 'layout') {
-          if (!subLayouts.includes(face[key]) || goingBackwards) {
-            faces[0][key] = face[key];
-          }
-        } else if (
-          ['adventure', 'prepare'].includes(face.layout) &&
-          ['cmc', 'colors'].includes(key)
-        ) {
-        } else if (key == 'maintype') {
-        } else if (key == 'cmc') {
-          if (!subLayouts.includes(face.layout)) {
-            faces[0][key] += face[key];
-          }
-        } else if (key == 'colors') {
-          if (!subLayouts.includes(face.layout)) {
-            if (faces[0].colors && face[key]) {
-              face[key].forEach(color => {
-                if (!faces[0].colors?.includes(color)) {
-                  faces[0].colors?.push(color);
-                }
-              });
-            } else if (face[key]) {
-              faces[0].colors = face[key];
-            }
-          }
-        } else if (key == 'picurl') {
-          if (!faces[0][key] && face[key]) {
-            faces[0][key] = face[key];
-          }
-        } else if (faces[0][key]) {
-          if (key !== 'text') {
-            const needed = i - ((faces[0][key] as string).match(/ \/\/ /g)?.length || 0);
-            if (needed > 0) {
-              faces[0][key] += ' // '.repeat(needed);
-            }
-          }
-          faces[0][key] += (key == 'text' ? '\n\n---\n\n' : ' // ') + face[key];
-        } else {
-          faces[0][key] = ' // '.repeat(i + 1) + face[key];
-        }
-      });
-    });
-    return faces[0];
-  };
+  // const mergeCockFaceProps = (faces: CockFaceProps[], goingBackwards?: boolean): CockFaceProps => {
+  //   faces.slice(1).forEach((face, i) => {
+  //     Object.keys(face).forEach(key => {
+  //       if (key == 'layout') {
+  //         if (!subLayouts.includes(face[key]) || goingBackwards) {
+  //           faces[0][key] = face[key];
+  //         }
+  //       } else if (
+  //         ['adventure', 'prepare'].includes(face.layout) &&
+  //         ['cmc', 'colors'].includes(key)
+  //       ) {
+  //       } else if (key == 'maintype') {
+  //       } else if (key == 'cmc') {
+  //         if (!subLayouts.includes(face.layout)) {
+  //           faces[0][key] += face[key];
+  //         }
+  //       } else if (key == 'colors') {
+  //         if (!subLayouts.includes(face.layout)) {
+  //           if (faces[0].colors && face[key]) {
+  //             face[key].forEach(color => {
+  //               if (!faces[0].colors?.includes(color)) {
+  //                 faces[0].colors?.push(color);
+  //               }
+  //             });
+  //           } else if (face[key]) {
+  //             faces[0].colors = face[key];
+  //           }
+  //         }
+  //       } else if (key == 'picurl') {
+  //         if (!faces[0][key] && face[key]) {
+  //           faces[0][key] = face[key];
+  //         }
+  //       } else if (faces[0][key]) {
+  //         if (key !== 'text') {
+  //           const needed = i - ((faces[0][key] as string).match(/ \/\/ /g)?.length || 0);
+  //           if (needed > 0) {
+  //             faces[0][key] += ' // '.repeat(needed);
+  //           }
+  //         }
+  //         faces[0][key] += (key == 'text' ? '\n\n---\n\n' : ' // ') + face[key];
+  //       } else {
+  //         faces[0][key] = ' // '.repeat(i + 1) + face[key];
+  //       }
+  //     });
+  //   });
+  //   return faces[0];
+  // };
   /**
    * Convert an hc all_parts array to a cockatrice related props array
    * @param all_parts hc all_parts array
@@ -253,12 +270,36 @@ export const toCockCube = ({
     });
     return cockRelateds;
   };
+  const compressHCCardFaces = (card: HCCard.Any) => {
+    if ('card_faces' in card) {      
+      const goingToCompressAll = Boolean(card.card_faces.length > 2 && card.card_faces.filter(face=>face.compress_face || face.drop_face).length == 1)
+      for (let i = card.card_faces.length - 1; i > 0; i--) {
+        if (i == 3 && card.card_faces[2].layout == 'transform') {
+          card.card_faces[i - 1] = mergeHCCardFaces([card.card_faces[i], card.card_faces[i-1]]);
+          card.card_faces.splice(i, 1);
+        } else if (card.card_faces[i].compress_face) {
+          card.card_faces[i - 1] = mergeHCCardFaces([card.card_faces[i - 1], card.card_faces[i]]);
+          card.card_faces.splice(i, 1);
+        } else if (card.card_faces[i].drop_face) {
+          card.card_faces.splice(i, 1);
+        }
+      }
+
+      // compress down to 1 side and use front image if there are still too many sides
+      if (goingToCompressAll || !card.card_faces[0].image) {
+        card.card_faces[0].image = card.image;
+      }
+
+    }
+    return card;
+  };
   /**
    * Convert an hc card to cockatrice props
    * @param card card to convert
    * @returns cockatrice props
    */
-  const hcCardToCockProps = (card: HCCard.Any): CockCardProps => {
+  const hcCardToCockProps = (uncompressedCard: HCCard.Any): CockCardProps => {
+    const card = compressHCCardFaces(uncompressedCard);
     const cockCard: CockCardProps = {
       coloridentity: card.color_identity.join(''),
       id: card.id,
@@ -281,62 +322,9 @@ export const toCockCube = ({
     } else {
       cockCard.props.push(hcFaceToCockProps(card));
     }
-    // compress/drop layouts that should always be compressed/dropped
-    if (cockCard.props.length > 1) {
-      for (let i = cockCard.props.length - 1; i > 0; i--) {
-        if (dropLayouts.includes(cockCard.props[i].layout)) {
-          cockCard.props.splice(i, 1);
-        } else if (alwaysCompressLayouts.includes(cockCard.props[i].layout)) {
-          cockCard.props[i - 1] = mergeCockFaceProps([cockCard.props[i - 1], cockCard.props[i]]);
-          cockCard.props.splice(i, 1);
-        }
-      }
-    }
-    // compress flips if necessary
-    if (cockCard.props.length > 2 && card.tags?.includes('flip')) {
-      for (let i = cockCard.props.length - 1; i > 0; i--) {
-        if (cockCard.props[i].layout == 'flip') {
-          if (i == 3 && cockCard.props[2].layout == 'transform') {
-            cockCard.props[i - 1] = mergeCockFaceProps(
-              [cockCard.props[i], cockCard.props[i - 1]],
-              true
-            );
-            cockCard.props.splice(i, 1);
-          } else {
-            cockCard.props[i - 1] = mergeCockFaceProps([cockCard.props[i - 1], cockCard.props[i]]);
-            cockCard.props.splice(i, 1);
-          }
-        }
-      }
-    }
-    // compress down to 1 side and use front image if there are still too many sides
-    if (cockCard.props.length > 2) {
-      cockCard.props = [mergeCockFaceProps(cockCard.props)];
-      cockCard.props[0].picurl = card.image;
-    } else if (!cockCard.props[0].picurl) {
-      cockCard.props[0].picurl = card.image;
-    }
     if (card.all_parts) {
       cockCard.related = hcAllPartsToCockRelated(card.all_parts);
     }
-    // make sure names aren't taken and then store the names
-    cockCard.props.forEach((face, i) => {
-      if (card.isActualToken) {
-        if (cockCard.props.length > 1) {
-          face.name += '1';
-        } else {
-          face.name = card.id;
-        }
-      }
-      if (!face.name) {
-        face.name = `(${i ? 'Back' : 'Front'} of ${cockCard.props[1 - i].name})`;
-      } else if (i && face.name == cockCard.props[0].name) {
-        face.name += ' (Back)';
-      }
-      while (nameIsTaken(face.name)) {
-        face.name += '_';
-      }
-    });
     idNames[cockCard.id] = cockCard.props.map(face => face.name);
     return cockCard;
   };
